SIG-3904 add triggers (for actions) to questionnaire data model
